+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - L2CAP: additional authorization_required param in l2cap_ecbm_register_service  
 - GATT Client: support GATT over Enhanced LE Bearer
 - GATT Server: support GATT over Enhanced LE Bearer
+- HOG Device: emit HIDS_SUBEVENT_SET_REPORT
 - LE Device DB: le_device_db_dump dumps LTK
 - Port for Zephyr 3.x